requirements.txt
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
googlesearch-python>=1.2.3
markdownify>=0.11.6
python-dotenv>=1.0.0
lxml>=4.9.0
html2text>=2020.1.16
tenacity>=8.2.0
